added swagger, some basic validation, tidied up loose edges.

+++ src/main/java/example/rabobank/toDo/service/ToDoServiceImpl.java
@@ -1,11 +1,10 @@
 package example.rabobank.toDo.service;
 
 import example.rabobank.toDo.entity.ToDo;
+import example.rabobank.toDo.exceptions.ToDoNotFoundException;
 import example.rabobank.toDo.repository.ToDoRepository;
-import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -20,9 +19,8 @@ public class ToDoServiceImpl implements ToDoService {
     @Override
     public ToDo getById(Long id) {
         return toDoRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("ToDo with id " + id + " not found"));
+                .orElseThrow(() -> new ToDoNotFoundException(id));
     }
-
     @Override
     public List<ToDo> getAll() {
         return toDoRepository.findAll();
@@ -32,33 +30,19 @@ public class ToDoServiceImpl implements ToDoService {
     public ToDo create(ToDo todo) {
         return toDoRepository.save(todo);
     }
-
     @Override
     public ToDo update(Long id, ToDo todo) {
         ToDo existingToDo = getById(id);
-
-        ToDo updatedToDo = ToDo.builder()
-                .id(existingToDo.getId())
-                .dateCreated(existingToDo.getDateCreated())
-                .description(todo.getDescription())
-                .completed(todo.isCompleted())
-                .deleted(todo.isDeleted())
-                .dateToComplete(todo.getDateToComplete())
-                .build();
-
-        return toDoRepository.save(updatedToDo);
+        existingToDo.updateWith(todo);
+        return toDoRepository.save(existingToDo);
     }
-
     @Override
     public void delete(Long id) {
+        if (!toDoRepository.existsById(id)) {
+            throw new ToDoNotFoundException(id);
+        }
         toDoRepository.deleteById(id);
     }
-
-    @Override
-    public List<ToDo> getByCreatedDate(LocalDate date) {
-        return toDoRepository.findByDateCreated(date);
-    }
-
     @Override
     public List<ToDo> getByCompleted() {
         return toDoRepository.findByCompleted(true);
